Cleanup...

+++ src/main/java/vazkii/botania/client/core/handler/TerraPickRankDisplayHandler.java
@@ -26,7 +26,6 @@ import net.minecraft.util.StatCollector;
 import org.lwjgl.input.Mouse;
 import org.lwjgl.opengl.GL11;
 
-import vazkii.botania.client.core.helper.RenderHelper;
 import vazkii.botania.common.item.equipment.tool.ItemTerraPick;
 import vazkii.botania.common.lib.LibObfuscation;
 import cpw.mods.fml.relauncher.ReflectionHelper;
@@ -51,11 +50,11 @@ public final class TerraPickRankDisplayHandler {
 					int width = 0;
 					for(String s : tooltip)
 						width = Math.max(width, font.getStringWidth(s));
-					
+
 					int offx = 11;
 					int offy = 17;
 					int height = 3;
-					
+
 					int level = ItemTerraPick.getLevel(stack);
 					int max = ItemTerraPick.LEVELS[Math.min(ItemTerraPick.LEVELS.length - 1, level + 1)];
 					boolean ss = level >= ItemTerraPick.LEVELS.length - 1;
@@ -63,14 +62,14 @@ public final class TerraPickRankDisplayHandler {
 					float percent = level == 0 ? 0F : (float) curr / (float) max;
 					int rainbowWidth = Math.min(width - (ss ? 0 : 1), (int) (width * percent));
 					float huePer = width == 0 ? 0F : 1F / width;
-					float hueOff = ((float) ClientTickHandler.ticksInGame + ClientTickHandler.partialTicks) * 0.01F;
-					
+					float hueOff = (ClientTickHandler.ticksInGame + ClientTickHandler.partialTicks) * 0.01F;
+
 					GL11.glDisable(GL11.GL_DEPTH_TEST);
 					Gui.drawRect(mouseX + offx - 1, mouseY - offy - height - 1, mouseX + offx + width + 1, mouseY - offy, 0xFF000000);
 					for(int i = 0; i < rainbowWidth; i++)
 						Gui.drawRect(mouseX + offx + i, mouseY - offy - height, mouseX + offx + i + 1, mouseY - offy, Color.HSBtoRGB(hueOff + huePer * i, 1F, 1F));
 					Gui.drawRect(mouseX + offx + rainbowWidth, mouseY - offy - height, mouseX + offx + width, mouseY - offy, 0xFF555555);
-					
+
 					String rank = StatCollector.translateToLocal("botania.rank" + level).replaceAll("&", "\u00a7");
 					boolean light = GL11.glGetBoolean(GL11.GL_LIGHTING);
 					GL11.glDisable(GL11.GL_LIGHTING);
@@ -87,5 +86,5 @@ public final class TerraPickRankDisplayHandler {
 			}
 		}
 	}
-	
+
 }
